requirements.txt
markdown
requests
python-docx
docx2pdf
tkcalendar